+++ Craftsman/Builders/Tests/Fakes/FakeEntityBuilderBuilder.cs
@@ -37,24 +37,26 @@ public class FakeEntityBuilderBuilder
     private static string GetCreateFakeBuilderFileText(string classNamespace, Entity entity, string srcDirectory, string testDirectory, string projectBaseName)
     {
         var entitiesClassPath = ClassPathHelper.EntityClassPath(testDirectory, "", entity.Plural, projectBaseName);
-        var dtoClassPath = ClassPathHelper.DtoClassPath(srcDirectory, "", entity.Plural, projectBaseName);
-        var creationDtoName = FileNames.GetDtoName(entity.Name, Dto.Creation);
-        var fakeCreationDtoName = $"Fake{creationDtoName}";
+        var modelClassPath = ClassPathHelper.EntityModelClassPath(srcDirectory, entity.Name, entity.Plural, null, projectBaseName);
+        var creationModelName = EntityModel.Creation.GetClassName(entity.Name);
+        var fakeCreationModelName = FileNames.FakerName(creationModelName);
+        
+        var propHelpers = EntityModelPropBuilder(FileNames.FakeBuilderName(entity.Name), entity.Properties);
 
         return @$"namespace {classNamespace};
 
 using {entitiesClassPath.ClassNamespace};
-using {dtoClassPath.ClassNamespace};
+using {modelClassPath.ClassNamespace};
 
 public class {FileNames.FakeBuilderName(entity.Name)}
 {{
-    private {creationDtoName} _creationData = new {fakeCreationDtoName}().Generate();
+    private {creationModelName} _creationData = new {fakeCreationModelName}().Generate();
 
-    public {FileNames.FakeBuilderName(entity.Name)} WithDto({creationDtoName} dto)
+    public {FileNames.FakeBuilderName(entity.Name)} WithModel({creationModelName} model)
     {{
-        _creationData = dto;
+        _creationData = model;
         return this;
-    }}
+    }}{propHelpers}
     
     public {entity.Name} Build()
     {{
@@ -62,5 +64,29 @@ public class {FileNames.FakeBuilderName(entity.Name)}
         return result;
     }}
 }}";
+    }
+
+    public static string EntityModelPropBuilder(string builderName, List<EntityProperty> props)
+    {
+        var propString = string.Empty;
+        for (var eachProp = 0; eachProp < props.Count; eachProp++)
+        {
+            if (!props[eachProp].CanManipulate)
+                continue;
+            if (props[eachProp].IsForeignKey && props[eachProp].IsMany)
+                continue;
+            if (!props[eachProp].IsPrimitiveType)
+                continue;
+
+            propString += $@"
+    
+    public {builderName} With{props[eachProp].Name}({props[eachProp].Type} {props[eachProp].Name.LowercaseFirstLetter()})
+    {{
+        _creationData.{props[eachProp].Name} = {props[eachProp].Name.LowercaseFirstLetter()};
+        return this;
+    }}";
+        }
+
+        return propString;
     }
 }