+++ Craftsman/Builders/Tests/FunctionalTests/AddListTestBuilder.cs
@@ -78,7 +78,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)} : TestB
     public async Task {testName}()
     {{
         // Arrange
-        var {fakeParentEntity} = Fake{feature.ParentEntity}.Generate(new {fakeParentCreationDto}().Generate());
+        var {fakeParentEntity} = new Fake{feature.ParentEntity}Builder().Build();
         await InsertAsync({fakeParentEntity});
         var {fakeEntityVariableName} = new List<{createDto}> {{new {fakeEntityForCreation} {{ }}.Generate()}};{clientAuth}