+++ Craftsman/Builders/Tests/IntegrationTests/AddCommandTestBuilder.cs
@@ -70,9 +70,8 @@ public class {classPath.ClassNameWithoutExt} : TestBase
                 var baseVarName = entityProperty.ForeignEntityName != entity.Name
                     ? $"{entityProperty.ForeignEntityName}"
                     : $"{entityProperty.ForeignEntityName}Parent";
-                var fakeParentClass = FileNames.FakerName(entityProperty.ForeignEntityName);
-                var fakeParentCreationDto = FileNames.FakerName(FileNames.GetDtoName(entityProperty.ForeignEntityName, Dto.Creation));
-                fakeParent += @$"var fake{baseVarName}One = {fakeParentClass}.Generate(new {fakeParentCreationDto}().Generate());
+                var fakeParentBuilder = FileNames.FakeBuilderName(entityProperty.ForeignEntityName);
+                fakeParent += @$"var fake{baseVarName}One = new {fakeParentBuilder}().Build();
         await testingServiceScope.InsertAsync(fake{baseVarName}One);{Environment.NewLine}{Environment.NewLine}        ";
                 fakeParentIdRuleFor +=
                     $"{Environment.NewLine}            .RuleFor({entity.Lambda} => {entity.Lambda}.{entityProperty.Name}, _ => fake{baseVarName}One.Id)";