+++ Craftsman/Builders/Tests/UnitTests/RolePermissionsUnitTestBuilder.cs
@@ -32,9 +32,9 @@ public class RolePermissionsUnitTestBuilder
         var wrapperClassPath = ClassPathHelper.WrappersClassPath(srcDirectory, "", projectBaseName);
         var domainPolicyClassPath = ClassPathHelper.PolicyDomainClassPath(srcDirectory, "", projectBaseName);
         var entityClassPath = ClassPathHelper.EntityClassPath(srcDirectory, "", "RolePermissions", projectBaseName);
-        var dtoClassPath = ClassPathHelper.DtoClassPath(srcDirectory, "", "RolePermissions", projectBaseName);
         var rolesClassPath = ClassPathHelper.EntityClassPath(srcDirectory, "", "Roles", projectBaseName);
         var errorsClassPath = ClassPathHelper.ExceptionsClassPath(solutionDirectory, "");
+        var modelClassPath = ClassPathHelper.EntityModelClassPath(srcDirectory, "RolePermission", "RolePermissions", null, projectBaseName);
 
         return @$"namespace {classPath.ClassNamespace};
 
@@ -42,7 +42,7 @@ using {errorsClassPath.ClassNamespace};
 using {domainPolicyClassPath.ClassNamespace};
 using {entityClassPath.ClassNamespace};
 using {wrapperClassPath.ClassNamespace};
-using {dtoClassPath.ClassNamespace};
+using {modelClassPath.ClassNamespace};
 using {rolesClassPath.ClassNamespace};
 using Bogus;
 using FluentAssertions;
@@ -66,7 +66,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
         var role = _faker.PickRandom(Role.ListNames());
 
         // Act
-        var newRolePermission = RolePermission.Create(new RolePermissionForCreationDto()
+        var newRolePermission = RolePermission.Create(new RolePermissionForCreation()
         {{
             Permission = permission,
             Role = role
@@ -81,7 +81,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void can_NOT_create_rolepermission_with_invalid_role()
     {{
         // Arrange
-        var rolePermission = () => RolePermission.Create(new RolePermissionForCreationDto()
+        var rolePermission = () => RolePermission.Create(new RolePermissionForCreation()
         {{
             Permission = _faker.PickRandom(Permissions.List()),
             Role = _faker.Lorem.Word()
@@ -95,7 +95,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void can_NOT_create_rolepermission_with_invalid_permission()
     {{
         // Arrange
-        var rolePermission = () => RolePermission.Create(new RolePermissionForCreationDto()
+        var rolePermission = () => RolePermission.Create(new RolePermissionForCreation()
         {{
             Role = _faker.PickRandom(Role.ListNames()),
             Permission = _faker.Lorem.Word()
@@ -112,16 +112,16 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
         var wrapperClassPath = ClassPathHelper.WrappersClassPath(srcDirectory, "", projectBaseName);
         var domainPolicyClassPath = ClassPathHelper.PolicyDomainClassPath(srcDirectory, "", projectBaseName);
         var entityClassPath = ClassPathHelper.EntityClassPath(srcDirectory, "", "RolePermissions", projectBaseName);
-        var dtoClassPath = ClassPathHelper.DtoClassPath(srcDirectory, "", "RolePermissions", projectBaseName);
         var rolesClassPath = ClassPathHelper.EntityClassPath(srcDirectory, "", "Roles", projectBaseName);
         var errorsClassPath = ClassPathHelper.ExceptionsClassPath(solutionDirectory, "");
+        var modelClassPath = ClassPathHelper.EntityModelClassPath(srcDirectory, "RolePermission", "RolePermissions", null, projectBaseName);
 
         return @$"namespace {classPath.ClassNamespace};
 
 using {domainPolicyClassPath.ClassNamespace};
 using {entityClassPath.ClassNamespace};
 using {wrapperClassPath.ClassNamespace};
-using {dtoClassPath.ClassNamespace};
+using {modelClassPath.ClassNamespace};
 using {rolesClassPath.ClassNamespace};
 using {errorsClassPath.ClassNamespace};
 using Bogus;
@@ -142,7 +142,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void can_update_rolepermission()
     {{
         // Arrange
-        var rolePermission = RolePermission.Create(new RolePermissionForCreationDto()
+        var rolePermission = RolePermission.Create(new RolePermissionForCreation()
         {{
             Permission = _faker.PickRandom(Permissions.List()),
             Role = _faker.PickRandom(Role.ListNames())
@@ -151,7 +151,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
         var role = _faker.PickRandom(Role.ListNames());
         
         // Act
-        rolePermission.Update(new RolePermissionForUpdateDto()
+        rolePermission.Update(new RolePermissionForUpdate()
         {{
             Permission = permission,
             Role = role
@@ -166,12 +166,12 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void can_NOT_update_rolepermission_with_invalid_role()
     {{
         // Arrange
-        var rolePermission = RolePermission.Create(new RolePermissionForCreationDto()
+        var rolePermission = RolePermission.Create(new RolePermissionForCreation()
         {{
             Permission = _faker.PickRandom(Permissions.List()),
             Role = _faker.PickRandom(Role.ListNames())
         }});
-        var updateRolePermission = () => rolePermission.Update(new RolePermissionForUpdateDto()
+        var updateRolePermission = () => rolePermission.Update(new RolePermissionForUpdate()
         {{
             Permission = _faker.PickRandom(Permissions.List()),
             Role = _faker.Lorem.Word()
@@ -185,12 +185,12 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void can_NOT_update_rolepermission_with_invalid_permission()
     {{
         // Arrange
-        var rolePermission = RolePermission.Create(new RolePermissionForCreationDto()
+        var rolePermission = RolePermission.Create(new RolePermissionForCreation()
         {{
             Permission = _faker.PickRandom(Permissions.List()),
             Role = _faker.PickRandom(Role.ListNames())
         }});
-        var updateRolePermission = () => rolePermission.Update(new RolePermissionForUpdateDto()
+        var updateRolePermission = () => rolePermission.Update(new RolePermissionForUpdate()
         {{
             Permission = _faker.Lorem.Word(),
             Role = _faker.PickRandom(Role.ListNames())