+++ Craftsman/Builders/Tests/UnitTests/UserUnitTestBuilder.cs
@@ -32,10 +32,10 @@ public class UserUnitTestBuilder
         var wrapperClassPath = ClassPathHelper.WrappersClassPath(srcDirectory, "", projectBaseName);
         var emailClassPath = ClassPathHelper.EntityClassPath(srcDirectory, "", "Emails", projectBaseName);
         var entityClassPath = ClassPathHelper.EntityClassPath(srcDirectory, "", "Users", projectBaseName);
-        var dtoClassPath = ClassPathHelper.DtoClassPath(srcDirectory, "", "Users", projectBaseName);
         var domainEventsClassPath = ClassPathHelper.DomainEventsClassPath(srcDirectory, "", "Users", projectBaseName);
         var fakerClassPath = ClassPathHelper.TestFakesClassPath(solutionDirectory, "", "User", projectBaseName);
         var errorsClassPath = ClassPathHelper.ExceptionsClassPath(srcDirectory, "");
+        var modelClassPath = ClassPathHelper.EntityModelClassPath(srcDirectory, "User", "Users", null, projectBaseName);
 
         return @$"namespace {classPath.ClassNamespace};
 
@@ -43,7 +43,7 @@ using {domainEventsClassPath.ClassNamespace};
 using {emailClassPath.ClassNamespace};
 using {entityClassPath.ClassNamespace};
 using {wrapperClassPath.ClassNamespace};
-using {dtoClassPath.ClassNamespace};
+using {modelClassPath.ClassNamespace};
 using {fakerClassPath.ClassNamespace};
 using {errorsClassPath.ClassNamespace};
 using Bogus;
@@ -64,7 +64,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void can_create_valid_user()
     {{
         // Arrange
-        var toCreate = new FakeUserForCreationDto().Generate();
+        var toCreate = new FakeUserForCreation().Generate();
 
         // Act
         var newUser = User.Create(toCreate);
@@ -81,7 +81,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void can_NOT_create_user_without_identifier()
     {{
         // Arrange
-        var toCreate = new FakeUserForCreationDto().Generate();
+        var toCreate = new FakeUserForCreation().Generate();
         toCreate.Identifier = null;
         var newUser = () => User.Create(toCreate);
 
@@ -93,7 +93,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void can_NOT_create_user_with_whitespace_identifier()
     {{
         // Arrange
-        var toCreate = new FakeUserForCreationDto().Generate();
+        var toCreate = new FakeUserForCreation().Generate();
         toCreate.Identifier = "" "";
         var newUser = () => User.Create(toCreate);
 
@@ -104,12 +104,15 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     [Fact]
     public void queue_domain_event_on_create()
     {{
-        // Arrange + Act
-        var fakeRecipe = FakeUser.Generate();
+        // Arrange
+        var toCreate = new FakeUserForCreation().Generate();
+
+        // Act
+        var newUser = User.Create(toCreate);
 
         // Assert
-        fakeRecipe.DomainEvents.Count.Should().Be(1);
-        fakeRecipe.DomainEvents.FirstOrDefault().Should().BeOfType(typeof(UserCreated));
+        newUser.DomainEvents.Count.Should().Be(1);
+        newUser.DomainEvents.FirstOrDefault().Should().BeOfType(typeof(UserCreated));
     }}
 }}";
     }
@@ -119,7 +122,7 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
         var wrapperClassPath = ClassPathHelper.WrappersClassPath(srcDirectory, "", projectBaseName);
         var domainPolicyClassPath = ClassPathHelper.PolicyDomainClassPath(srcDirectory, "", projectBaseName);
         var entityClassPath = ClassPathHelper.EntityClassPath(srcDirectory, "", "Users", projectBaseName);
-        var dtoClassPath = ClassPathHelper.DtoClassPath(srcDirectory, "", "Users", projectBaseName);
+        var modelClassPath = ClassPathHelper.EntityModelClassPath(srcDirectory, "User", "Users", null, projectBaseName);
         var domainEventsClassPath = ClassPathHelper.DomainEventsClassPath(srcDirectory, "", "Users", projectBaseName);
         var fakerClassPath = ClassPathHelper.TestFakesClassPath(solutionDirectory, "", "User", projectBaseName);
         var errorsClassPath = ClassPathHelper.ExceptionsClassPath(srcDirectory, "");
@@ -130,7 +133,7 @@ using {domainEventsClassPath.ClassNamespace};
 using {domainPolicyClassPath.ClassNamespace};
 using {entityClassPath.ClassNamespace};
 using {wrapperClassPath.ClassNamespace};
-using {dtoClassPath.ClassNamespace};
+using {modelClassPath.ClassNamespace};
 using {fakerClassPath.ClassNamespace};
 using {errorsClassPath.ClassNamespace};
 using Bogus;
@@ -151,8 +154,8 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void can_update_user()
     {{
         // Arrange
-        var fakeUser = FakeUser.Generate();
-        var updatedUser = new FakeUserForUpdateDto().Generate();
+        var fakeUser = new FakeUserBuilder().Build();
+        var updatedUser = new FakeUserForUpdate().Generate();
         
         // Act
         fakeUser.Update(updatedUser);
@@ -169,8 +172,8 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void can_NOT_update_user_without_identifier()
     {{
         // Arrange
-        var fakeUser = FakeUser.Generate();
-        var updatedUser = new FakeUserForUpdateDto().Generate();
+        var fakeUser = new FakeUserBuilder().Build();
+        var updatedUser = new FakeUserForUpdate().Generate();
         updatedUser.Identifier = null;
         var newUser = () => fakeUser.Update(updatedUser);
 
@@ -182,8 +185,8 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void can_NOT_update_user_with_whitespace_identifier()
     {{
         // Arrange
-        var fakeUser = FakeUser.Generate();
-        var updatedUser = new FakeUserForUpdateDto().Generate();
+        var fakeUser = new FakeUserBuilder().Build();
+        var updatedUser = new FakeUserForUpdate().Generate();
         updatedUser.Identifier = "" "";
         var newUser = () => fakeUser.Update(updatedUser);
 
@@ -195,8 +198,8 @@ public class {Path.GetFileNameWithoutExtension(classPath.FullClassPath)}
     public void queue_domain_event_on_update()
     {{
         // Arrange
-        var fakeUser = FakeUser.Generate();
-        var updatedUser = new FakeUserForUpdateDto().Generate();
+        var fakeUser = new FakeUserBuilder().Build();
+        var updatedUser = new FakeUserForUpdate().Generate();
         fakeUser.DomainEvents.Clear();
         
         // Act